Create roundTheGrades.php
This problem is all about rounding the student grades from the grades array   if the nearest multiply of 5 is at a distance of less than 3. 

I mean: if 68 then it will be 70.
             if 73 it will be 75 .
             if 38 then 40. 
     
if 67 then it remains the same as 70-67 is not less than 3. 

and under 40 scores will not be counted at the rounding value.